app/requirements.txt
astroid==2.4.1
Authlib==0.14.3
azure-common==1.1.25
azure-nspkg==3.0.2
azure-storage==0.36.0
cachelib==0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dominate==2.5.1
Flask==1.1.2
flask-azure-oauth==0.3.0
Flask-Bootstrap==3.3.7.1
Flask-Session==0.3.2
Flask-WTF==0.14.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msal==1.3.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
